Simplify grid nesting
