+++ client-vue/src/views/HomePage.vue
@@ -1,16 +1,47 @@
 <template>
-  <v-container>
-    <LocationMap />
-  </v-container>
+  <div>
+    <v-toolbar flat>
+      <v-toolbar-title>Location Map</v-toolbar-title>
+      <v-spacer></v-spacer>
+      <div>
+        <v-switch v-model="sticky" hide-details label="Auto-select and print"></v-switch>
+      </div>
+      <v-btn class="ml-3" color="primary" type="button" @click="createPrinterModal()">
+        Create Printer
+      </v-btn>
+    </v-toolbar>
+
+    <v-container>
+      <PrinterGrid />
+    </v-container>
+
+    <CreatePrinterDialog :show.sync="showDialog" v-on:update:show="onChangeShowDialog($event)" />
+  </div>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
-import LocationMap from "../components/PrinterGrid/LocationMap.vue";
 import { Component } from "vue-property-decorator";
+import CreatePrinterDialog from "@/components/Dialogs/CreatePrinterDialog.vue";
+import PrinterGrid from "@/components/PrinterGrid/PrinterGrid.vue";
 
 @Component({
-  components: { LocationMap }
+  data: () => ({
+    sticky: false
+  }),
+  components: { PrinterGrid, CreatePrinterDialog }
 })
-export default class HomePage extends Vue {}
+export default class HomePage extends Vue {
+  sticky: false;
+  showDialog = false;
+
+  async createPrinterModal() {
+    this.showDialog = true;
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  onChangeShowDialog(event: any) {
+    // Placeholder
+  }
+}
 </script>
